Crate curve_arithmetic

Basic definitions of the curve and pairing abstractions, and implementations of these abstractions for the curves used on Concordium.

Re-exports

pub use secret_value::Secret;

pub use secret_value::Value;

Modules

secret_value
A thin wrapper around a scalar to indicate that it is a secret value.

Enums

CurveDecodingError

Traits

Curve
A relatively large trait that covers what is needed to perform constructions and proofs upon a base group. This can only be implemented by groups of prime order size. More correctly this would be called a group, since it is generally a subset of an elliptic curve, but the name is in use now.

Pairing
A pairing friendly curve is a collection of two groups and a pairing function. The groups must be of prime order.

Functions

multiexp
Like ‘multiexp_worker’, but computes a reasonable window size automatically.

multiexp_table
Compute the table of powers that can be used multiexp_worker_given_table.

multiexp_worker
This implements the WNAF method from https://link.springer.com/content/pdf/10.1007%2F3-540-45537-X_13.pdf

multiexp_worker_given_table
This function assumes the same properties about the inputs as multiexp_worker, as well as the fact that the table corresponds to the window-size and the given inputs.