Module curve_arithmetic::secret_value

A thin wrapper around a scalar to indicate that it is a secret value.

Structs

Secret

A generic wrapper for a secret that implements a zeroize on drop. Other types are expected to wrap this in more convenient interfaces. Ideally the constraint would be Default, but fields we have do not implement it, so we cannot use it at the moment. Hence the temporary hack of ‘F: Field’.

Value

A secret value. The idea of this datatype is to mark some scalars as secret, so that their use is harder and there is no implicit copy.