Trait curve_arithmetic::Curve

pub trait Curve: Serialize + Copy + Clone + Sized + Send + Sync + Debug + Display + PartialEq + Eq + 'static {
    type Scalar: PrimeField + Field + Serialize;
    type Base: Field;
    type Compressed;

    const SCALAR_LENGTH: usize;
    const GROUP_ELEMENT_LENGTH: usize;

    fn zero_point() -> Self;

    fn one_point() -> Self;

    fn is_zero_point(&self) -> bool;

    fn inverse_point(&self) -> Self;

    fn double_point(&self) -> Self;

    fn plus_point(&self, other: &Self) -> Self;

    fn minus_point(&self, other: &Self) -> Self;

    fn mul_by_scalar(&self, scalar: &Self::Scalar) -> Self;

    fn compress(&self) -> Self::Compressed;

    fn decompress(c: &Self::Compressed) -> Result<Self, CurveDecodingError>;

    fn decompress_unchecked(
        c: &Self::Compressed
    ) -> Result<Self, CurveDecodingError>;

    fn bytes_to_curve_unchecked<R: ReadBytesExt>(b: &mut R) -> Result<Self>;

    fn generate<R: Rng>(rng: &mut R) -> Self;

    fn generate_scalar<R: Rng>(rng: &mut R) -> Self::Scalar;

    fn scalar_from_u64(n: u64) -> Self::Scalar;

    fn scalar_from_bytes<A: AsRef<[u8]>>(bs: A) -> Self::Scalar;

    fn hash_to_group(m: &[u8]) -> Self;

    fn generate_non_zero_scalar<R: Rng>(rng: &mut R) -> Self::Scalar { ... }

}

A relatively large trait that covers what is needed to perform constructions and proofs upon a base group. This can only be implemented by groups of prime order size. More correctly this would be called a group, since it is generally a subset of an elliptic curve, but the name is in use now.

Associated Types

type Scalar: PrimeField + Field + Serialize

The prime field of the group order size.

type Base: Field

The base field of the curve. In general larger than the Scalar field.

type Compressed

A compressed representation of curve points used for compact serialization.

Associated Constants

const SCALAR_LENGTH: usize

Size in bytes of elements of the Curve::Scalar field.

const GROUP_ELEMENT_LENGTH: usize

Size in bytes of group elements when serialized.

Required methods

fn zero_point() -> Self

Unit for the group operation.

fn one_point() -> Self

Chosen generator of the group.

fn is_zero_point(&self) -> bool

fn inverse_point(&self) -> Self

Return the group inverse of the given element.

fn double_point(&self) -> Self

Given x compute x + x.

fn plus_point(&self, other: &Self) -> Self

The group operation.

fn minus_point(&self, other: &Self) -> Self

Subtraction. This is generally more efficient than a combination of Curve::inverse_point and Curve::plus_point.

fn mul_by_scalar(&self, scalar: &Self::Scalar) -> Self

Exponentiation by a scalar, i.e., compute n * x for a group element x and integer n.

fn compress(&self) -> Self::Compressed

fn decompress(c: &Self::Compressed) -> Result<Self, CurveDecodingError>

fn decompress_unchecked(
c: &Self::Compressed
) -> Result<Self, CurveDecodingError>

fn bytes_to_curve_unchecked<R: ReadBytesExt>(b: &mut R) -> Result<Self>

Deserialize a value from a byte source, but do not check that it is in the group itself. This can be cheaper if the source of the value is trusted, but it must not be used on untrusted sources.

fn generate<R: Rng>(rng: &mut R) -> Self

Generate a random group element, uniformly distributed.

fn generate_scalar<R: Rng>(rng: &mut R) -> Self::Scalar

Generate a random scalar value, uniformly distributed.

fn scalar_from_u64(n: u64) -> Self::Scalar

Make a scalar from a 64-bit unsigned integer. This function assumes that the field is big enough to accommodate any 64-bit unsigned integer.

fn scalar_from_bytes<A: AsRef<[u8]>>(bs: A) -> Self::Scalar

Make a scalar by taking the first Scalar::CAPACITY bits and interpreting them as a little-endian integer.

fn hash_to_group(m: &[u8]) -> Self

Hash to a curve point from a seed. This is deterministic function.

Provided methods

fn generate_non_zero_scalar<R: Rng>(rng: &mut R) -> Self::Scalar

Generate a non-zero scalar. The default implementation does repeated sampling until a non-zero scalar is reached.