Trait curve_arithmetic::Pairing

pub trait Pairing: Sized + 'static + Clone {
    type ScalarField: PrimeField + Serialize;
    type G1: Curve<Base = Self::BaseField, Scalar = Self::ScalarField>;
    type G2: Curve<Base = Self::BaseField, Scalar = Self::ScalarField>;
    type G1Prepared;
    type G2Prepared;
    type BaseField: PrimeField;
    type TargetField: Field + Serial;
    fn miller_loop<'a, I>(i: I) -> Self::TargetField
    where
        I: IntoIterator<Item = &'a (&'a Self::G1Prepared, &'a Self::G2Prepared)>;

    fn final_exponentiation(_: &Self::TargetField) -> Option<Self::TargetField>;

    fn g1_prepare(_: &Self::G1) -> Self::G1Prepared;

    fn g2_prepare(_: &Self::G2) -> Self::G2Prepared;

    fn generate_scalar<R: Rng>(rng: &mut R) -> Self::ScalarField;

    fn check_pairing_eq(
        g1x: &Self::G1, 
        g2x: &Self::G2, 
        g1y: &Self::G1, 
        g2y: &Self::G2
    ) -> bool { ... }

    fn pairing_product(
        g1x: &Self::G1, 
        g2x: &Self::G2, 
        g1y: &Self::G1, 
        g2y: &Self::G2
    ) -> Option<Self::TargetField> { ... }

    fn pair(p: &Self::G1, q: &Self::G2) -> Self::TargetField { ... }

    fn generate_non_zero_scalar<R: Rng>(rng: &mut R) -> Self::ScalarField { ... }
}

A pairing friendly curve is a collection of two groups and a pairing function. The groups must be of prime order.

Associated Types

type ScalarField: PrimeField + Serialize

type G1: Curve<Base = Self::BaseField, Scalar = Self::ScalarField>

The first group of the pairing.

type G2: Curve<Base = Self::BaseField, Scalar = Self::ScalarField>

The second group, must have the same order as Pairing::G1. Both G1 and G2 must be of prime order size.

type G1Prepared

An auxiliary type that is used as an input to the pairing function.

type G2Prepared

An auxiliary type that is used as an input to the pairing function.

type BaseField: PrimeField

Field of the size of G1 and G2.

type TargetField: Field + Serial

The target of the pairing function. The pairing function actually maps to a subgroup of the same order as G1 and G2, but this subgroup is not exposed here and is generally not useful. It is subgroup of the multiplicative subgroup of the field.

Loading content...
Required methods

fn miller_loop<'a, I>(i: I) -> Self::TargetField where
I: IntoIterator<Item = &'a (&'a Self::G1Prepared, &'a Self::G2Prepared)>,

Compute the miller loop on the given sequence of prepared points.

fn final_exponentiation(_: &Self::TargetField) -> Option<Self::TargetField>

fn g1_prepare(_: &Self::G1) -> Self::G1Prepared

fn g2_prepare(_: &Self::G2) -> Self::G2Prepared

fn generate_scalar<R: Rng>(rng: &mut R) -> Self::ScalarField

Loading content...
Provided methods

fn check_pairing_eq(
g1x: &Self::G1,
g2x: &Self::G2,
g1y: &Self::G1,
g2y: &Self::G2
) -> bool

Check whether the pairing equation holds given the left and right-hand sides.

fn pairing_product(
g1x: &Self::G1,
g2x: &Self::G2,
g1y: &Self::G1,
g2y: &Self::G2
) -> Option<Self::TargetField>

Compute the product of the pairings, but more efficiently.

fn pair(p: &Self::G1, q: &Self::G2) -> Self::TargetField

fn generate_non_zero_scalar<R: Rng>(rng: &mut R) -> Self::ScalarField

Generate non-zero scalar by repeated sampling. Can be overriden by a more efficient implementation.

Loading content...